[RISC-V][HWASAN] Change the tag length to 7 bits
The RISC-V HWASAN implementation was merged two years before the ISA
specification was frozen, and its use of 8-bit tags no longer matches
the tag lengths supported by final ISA extension (7 or 16 bits). Update
the HWASAN implementation to use 7 bit tags, as those are compatible
with Sv57 (paging using 57-bit virtual addresses).

Remove the use of the "_v2" suffix, as that is leftover from copying the
AArch64 implementation, and provides an obvious compatibility break with
the pre-standard version of this code.